FILE: renames to-array to to-ids
